feat: ブラウザバックを無効化する処理をカスタムフックにし、それを適用。
